+++ spec/fixtures/code_to_test_ratio/app/models/dummy.rb
@@ -1,0 +1,9 @@
+this
+is
+lame
+this
+is
+lame
+this
+is
+lame